first refactor
theoretically this should all work, but i'm not entirely sure. we hnavent gotten any of the physical in bits in place for testing, but i'm assuming it should all work. let's pray for the bes.t

and since it feels like it, lets use this section as a personal diary entry.

right now, it's saturday. it's less than a week before competition, so we're very close to being done with robotics, forever. of course, i'll make sure to work on robotics in university. but asside from that, i feel weird. almost like a preemptive happiness for things to be over. i'm not sure what it is. i guess i'm also excited about my job; i really cant wait to start. i have this one math test on tuesday, but then i should be done with everything i have aside from that crappy english project. man, the timeline for that is gonna be ROUGH. i really dont want to work on it during tomorrow, but i guess i have to. anyways, gary just got back so i guess it's time to get back to work.. time will tell.....